# Order-Comments-Api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
        "/api/updateAllExpectedShipDates": {
            "get": {
                "tags": [
                    "UpdateAllExpectedShipDates"
                ],
                "summary": "Parse through all the comment fields in the MySql table to find a expected ship dates.",
                "operationId": "2dc84be86c31c0ad870dec4f1afe9c51",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/updateExpectedShipDate/{orderId}": {
            "get": {
                "tags": [
                    "UpdateExpectedShipDate"
                ],
                "summary": "Parse through a comment field in the MySql table to find a expected ship date.",
                "operationId": "372fd3fc1b2e5de8f132f89efd0cde77",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "path",
                        "description": "ID of Order",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/getSortedComments": {
            "get": {
                "tags": [
                    "GetSortedComments"
                ],
                "summary": "Get sorted comments from my MySql.",
                "operationId": "ec7bedcdb8a55f671a2e336988e25508",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/createOrder": {
            "post": {
                "tags": [
                    "CreateOrder"
                ],
                "summary": "Create an order.",
                "operationId": "2f32b864b16a8834cef2645266e4b8ec",
                "parameters": [
                    {
                        "name": "Comment",
                        "in": "query",
                        "description": "Order Comment.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/updateOrder/{orderId}": {
            "post": {
                "tags": [
                    "UpdateOrder"
                ],
                "summary": "Update and existing order.",
                "operationId": "9b75a2ba5e628dcd05f0c6cfecb222b2",
                "parameters": [
                    {
                        "name": "comments",
                        "in": "query",
                        "description": "Order Comment.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "orderId",
                        "in": "path",
                        "description": "ID of Order",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    }
}